Green do_not_delete_occupancy_rate_when_cocod_is_empty

+++ src/Domain/UseCases/OccupancyRate/DeleteOccupancyRate.cs
@@ -7,6 +7,9 @@ public class DeleteOccupancyRate {
     }
 
     public async Task Execute(string cocod) {
+        if (string.IsNullOrWhiteSpace(cocod)) {
+            throw new ArgumentException("Code is required");
+        }
         await availabilitySynchronizerApiClient.DeleteOccupancyRate(cocod);
     }
 }
